stop using parse_line from Text::ParseWords...
I reluctantly had to go back to a simplistic 'split()' with a negative
look-behind to handle the simplest cases of an escaped ';'.

The module would have been nice, but can't do what I want.  When the
input is, say:

    abc; /\.foo/\; echo $HOME; ghi

I want it to be parsed as the following commands:

    abc
    /\.foo/; echo $HOME
    ghi

Notice I want the "\" retained in the regex, but not the one escaping
the semicolon.

But parse_line will either lose both the backslashes or retain both of
them; you can't tell it to lose them only on the delimiter used.  Pity.

You could argue that the correct solution is to let it lose all the
delimiters and make the user add more.  This is certainly what *every*
tool does, but it increases clutter in what I want to be a really really
small language -- saying /\\\\.foo/ just doesn't have the same charm as
/\.foo/ :)