use fixed timestamps (for predictable SHAs) only if running under harness
I'd like to be able to use tsh under normal conditions also, like just
saying 'tsh empty' to get an empty commit, and in that case I'd like the
commit to have the current time.

When not under a harness:

  - $tick does not get defined, nor do GIT_COMMITTER_DATE and
    GIT_AUTHOR_DATE.

  - 'tc' (test-commit) uses the current time instead of our own internal
    epoch of 1310000000.

  - 'tt' (test-tick) will 'sleep 1' to make the actual time go forward.

Note that harness commits move forward by 1 minute on a 'tt', but if we
change that now it'll affect a lot of SHAs in a lot of test scripts.